speed up air rebuilds
TailwindCSS is slow. Instead of running it in `just main-air`, configure
air to monitor the CSS output directory and rebuild+restart the server
if the output CSS has changed. `just tw` will need to be invoked
manually but this should be very infrequent.

From local testing this reduces the cycle time from 2+ secs to <1 sec.

Also ignore node_packages for now.